extract remaining git sys calls into seperate file
