+++ src/onefetch/info.rs
@@ -1,10 +1,9 @@
 use {
     crate::onefetch::{
-        cli::Cli, cli_utils, commit_info::CommitInfo, deps, error::*, language::Language,
-        license::Detector, repo::Repo, text_color::TextColor,
+        cli::Cli, cli_utils, commit_info::CommitInfo, deps, error::*, git_utils,
+        language::Language, license::Detector, repo::Repo, text_color::TextColor,
     },
     colored::{Color, ColoredString, Colorize},
-    std::process::Command,
 };
 
 pub struct Info {
@@ -219,13 +218,13 @@ impl Info {
         let git_username = repo.get_git_username()?;
         let number_of_tags = repo.get_number_of_tags()?;
         let number_of_branches = repo.get_number_of_branches()?;
-        let git_history = Info::get_git_history(&workdir, config.no_merges);
-        let creation_date = Info::get_creation_date(&git_history)?;
-        let number_of_commits = Info::get_number_of_commits(&git_history);
-        let authors = Info::get_authors(&git_history, config.number_of_authors);
-        let last_change = Info::get_date_of_last_commit(&git_history)?;
+        let git_history = git_utils::get_git_history(&workdir, config.no_merges)?;
+        let creation_date = git_utils::get_creation_date(&git_history)?;
+        let number_of_commits = git_utils::get_number_of_commits(&git_history);
+        let authors = git_utils::get_authors(&git_history, config.number_of_authors);
+        let last_change = git_utils::get_date_of_last_commit(&git_history)?;
         let git_version = cli_utils::get_git_version()?;
-        let repo_size = Info::get_packed_size(&workdir)?;
+        let repo_size = git_utils::get_packed_size(&workdir)?;
         let license = Detector::new()?.get_license(&workdir)?;
         let dependencies = deps::DependencyDetector::new().get_dependencies(&workdir)?;
         let (languages, lines_of_code) =
@@ -263,124 +262,6 @@ impl Info {
             text_colors,
             config,
         })
-    }
-
-    fn get_git_history(dir: &str, no_merges: bool) -> Vec<String> {
-        let mut args = vec!["-C", dir, "log"];
-        if no_merges {
-            args.push("--no-merges");
-        }
-
-        args.push("--pretty=%cr\t%ae\t%an");
-
-        let output = Command::new("git").args(args).output().expect("Failed to execute git.");
-
-        let output = String::from_utf8_lossy(&output.stdout);
-        output.lines().map(|x| x.to_string()).collect::<Vec<_>>()
-    }
-
-    fn get_authors(git_history: &[String], n: usize) -> Vec<(String, usize, usize)> {
-        let mut authors = std::collections::HashMap::new();
-        let mut author_name_by_email = std::collections::HashMap::new();
-        let mut total_commits = 0;
-        for line in git_history {
-            let author_email = line.split('\t').collect::<Vec<_>>()[1].to_string();
-            let author_name = line.split('\t').collect::<Vec<_>>()[2].to_string();
-            let commit_count = authors.entry(author_email.to_string()).or_insert(0);
-            author_name_by_email.entry(author_email.to_string()).or_insert(author_name);
-            *commit_count += 1;
-            total_commits += 1;
-        }
-
-        let mut authors: Vec<(String, usize)> = authors.into_iter().collect();
-        authors.sort_by(|(_, a_count), (_, b_count)| b_count.cmp(a_count));
-
-        authors.truncate(n);
-
-        let authors: Vec<(String, usize, usize)> = authors
-            .into_iter()
-            .map(|(author, count)| {
-                (
-                    author_name_by_email.get(&author).unwrap().trim_matches('\'').to_string(),
-                    count,
-                    count * 100 / total_commits,
-                )
-            })
-            .collect();
-
-        authors
-    }
-
-    fn get_date_of_last_commit(git_history: &[String]) -> Result<String> {
-        let last_commit = git_history.first();
-
-        let output = match last_commit {
-            Some(date) => date.split('\t').collect::<Vec<_>>()[0].to_string(),
-            None => "??".into(),
-        };
-
-        Ok(output)
-    }
-
-    fn get_creation_date(git_history: &[String]) -> Result<String> {
-        let first_commit = git_history.last();
-
-        let output = match first_commit {
-            Some(creation_time) => creation_time.split('\t').collect::<Vec<_>>()[0].to_string(),
-            None => "??".into(),
-        };
-
-        Ok(output)
-    }
-
-    fn get_number_of_commits(git_history: &[String]) -> String {
-        let number_of_commits = git_history.len();
-        number_of_commits.to_string()
-    }
-
-    fn get_packed_size(dir: &str) -> Result<String> {
-        let output = Command::new("git")
-            .arg("-C")
-            .arg(dir)
-            .arg("count-objects")
-            .arg("-vH")
-            .output()
-            .expect("Failed to execute git.");
-
-        let output = String::from_utf8_lossy(&output.stdout);
-        let lines = output.to_string();
-        let size_line = lines.split('\n').find(|line| line.starts_with("size-pack:"));
-
-        let repo_size = match size_line {
-            None => "??",
-            Some(size_str) => &(size_str[11..]),
-        };
-
-        let output = Command::new("git")
-            .arg("-C")
-            .arg(dir)
-            .arg("ls-files")
-            .output()
-            .expect("Failed to execute git.");
-        // To check if command executed successfully or not
-        let error = &output.stderr;
-
-        if error.is_empty() {
-            let output = String::from_utf8_lossy(&output.stdout);
-
-            let lines = output.to_string();
-            let files_list = lines.split('\n');
-            let mut files_count: u128 = 0;
-            for _file in files_list {
-                files_count += 1;
-            }
-            files_count -= 1; // As splitting giving one line extra(blank).
-            let res = repo_size.to_owned() + (" (") + &(files_count.to_string()) + (" files)");
-            Ok(res)
-        } else {
-            let res = repo_size;
-            Ok(res.into())
-        }
     }
 
     fn get_formatted_subtitle_label(&self, label: &str) -> ColoredString {
