+++ src/onefetch/mod.rs
@@ -1,15 +1,16 @@
-pub mod ascii_art;
+mod ascii_art;
 pub mod cli;
 pub mod cli_utils;
-pub mod commit_info;
-pub mod deps;
+mod commit_info;
+mod deps;
 pub mod error;
-pub mod image_backends;
+mod git_utils;
+mod image_backends;
 pub mod info;
-pub mod info_fields;
-pub mod language;
-pub mod license;
+mod info_fields;
+mod language;
+mod license;
 pub mod printer;
 pub mod repo;
-pub mod text_color;
+mod text_color;
 mod utils;
